+++ NEWS.md
@@ -26,6 +26,7 @@
 * `plot_model()` now supports `wbm()`-models from the *panelr*-package.
 * `sjt.itemanalysis()` now works on ordered factors. A clearer error message was added when unordered factors are used. The old error message was not helpful.
 * `sjp.stackfrq()` (now named: `plot_stackfrq()`) gets a `show.n`-argument to also show count values. This option can be combined with `show.prc`.
+* `sjp.stackfrq()` (now named: `plot_stackfrq()`) now also works on grouped data frames.
 
 ## Bug fixes